+++ src/main/java/com/example/demo/dto/store/StoreDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class StoreDto {
     private String product;
-    private Long quantity;
+    private double quantity;
     private double price;
     private String type;
     private boolean received;
