+++ src/main/java/com/example/demo/entity/ProductsReceipt.java
@@ -18,7 +18,7 @@ public class ProductsReceipt extends AbstractEntity {
     @Column(nullable = false)
     private String product;
     @Column(nullable = false)
-    private Long quantity;
+    private double quantity;
     private double price;
     private String type;
 }
